Inheriting from shared controller means inherited templates too
