--- app/views/comparisons/electricity_targets/_table.csv.ruby
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-render partial: 'comparisons/shared/targets/table'
