--- app/views/comparisons/gas_targets/_table.html.erb
@@ -1,1 +1,0 @@
-<%= render partial: 'comparisons/shared/targets/table', locals: local_assigns %>
